Temporary location for StaticDependencyInjection
